params costs
