+++ GPTifier/include/cli.hpp
@@ -10,7 +10,6 @@ ParamsRun get_opts_run(int argc, char **argv);
 ParamsEmbedding get_opts_embed(int argc, char **argv);
 ParamsFineTune get_opts_create_fine_tuning_job(int argc, char **argv);
 ParamsGetFineTuningJobs get_opts_get_fine_tuning_jobs(int argc, char **argv);
-ParamsCosts get_opts_get_costs(int argc, char **argv);
 ParamsGetChatCompletions get_opts_get_chat_completions(int argc, char **argv);
 
 bool get_opts_files_list(int argc, char **argv);
